requirements.txt
flask
influxdb-client
python-dotenv
cloudinary
python-dateutil